Implement bodies as shared pointers
Celestial bodies will be often referred by different objects. Or they might be deleted in case of collision etc. Therefore, they are implemented in the simulation as shared pointers.
Defining them as such allows us to pass week_ptr of the body to many instances (or shared_ptr if required). Doing so prevent dangling pointers problems !

+ Added copy constructor for body

~ Bodies are now stacked in a shared pointer vector
~ Some code refactoring to accomodate the changes