Modify collision handling
To keep as much trails and datas up to date has possible, the collision handling model has been rework. Now  when a collision happen, the heaviest body (by mass) gets to absorb the other one when their mass ratio is under a constant.
For now the constant value is 1.01 so all collisions are "absorptions"

+ Added way for body to absorb others
~ Cleaned Body.cpp and Body.hpp